src/backend/requirements.txt
mpmath==1.3.0
networkx==3.5
numpy==2.3.3
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
polars==1.33.1
psutil==7.1.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.5
scikit-image==0.25.2
scipy==1.16.2
setuptools==80.9.0
six==1.17.0
sympy==1.14.0
tifffile==2025.9.20
torch==2.8.0
torchvision==0.23.0
typing_extensions==4.15.0
ultralytics==8.3.203
ultralytics-thop==2.0.17
urllib3==2.5.0
Werkzeug==3.1.3